Fix pidfile removal
The http://svn.middleware.georgetown.edu/view/cpp-sp?view=rev&revision=3209
upstream fix for https://bugs.internet2.edu/jira/browse/SSPCPP-272.

+++ shibd/shibd.cpp
@@ -387,7 +387,7 @@ int main(int argc, char *argv[])
             fprintf(stderr, "listener failure during service\n");
             listener->term();
             conf.term();
-            if (pidfile)
+            if (daemonize && pidfile)
                 unlink(pidfile);
             return -3;
         }
@@ -395,7 +395,7 @@ int main(int argc, char *argv[])
     }
 
     conf.term();
-    if (pidfile)
+    if (daemonize && pidfile)
         unlink(pidfile);
     return 0;
 }
